add task 3 and move checksum to the right folder
